+++ srcs/mixed_operators/matrix_operators.cpp
@@ -17,11 +17,11 @@ Matrix operator * (Matrix lhs, RationalNumber rhs)
 	const std::vector<std::vector<BaseAssignmentType *>> matrix = lhs.get_matrix();
 
 	std::vector<std::vector<BaseAssignmentType *>> matrix_res;
-	for (size_t i = 0; i < lhs.get_num_rows(); ++i)
+	for (int i = 0; i < lhs.get_num_rows(); ++i)
 	{
 		std::vector<BaseAssignmentType *> row;
 
-		for (size_t j = 0; j < lhs.get_num_cols(); j++)
+		for (int j = 0; j < lhs.get_num_cols(); j++)
 		{
 			BaseAssignmentType *res_term = matrix[i][j]->mult(&rhs);
 			if (res_term == nullptr) throw Ft_error("Invalid operation");
@@ -39,11 +39,11 @@ Matrix operator * (Matrix lhs, ImaginaryNumber rhs)
 	const std::vector<std::vector<BaseAssignmentType *>> matrix = lhs.get_matrix();
 
 	std::vector<std::vector<BaseAssignmentType *>> matrix_res;
-	for (size_t i = 0; i < lhs.get_num_rows(); ++i)
+	for (int i = 0; i < lhs.get_num_rows(); ++i)
 	{
 		std::vector<BaseAssignmentType *> row;
 
-		for (size_t j = 0; j < lhs.get_num_cols(); j++)
+		for (int j = 0; j < lhs.get_num_cols(); j++)
 		{
 			BaseAssignmentType *res_term = matrix[i][j]->mult(&rhs);
 			if (res_term == nullptr) throw Ft_error("Invalid operation");
@@ -61,11 +61,11 @@ Matrix operator / (Matrix lhs, RationalNumber rhs)
 	const std::vector<std::vector<BaseAssignmentType *>> matrix = lhs.get_matrix();
 
 	std::vector<std::vector<BaseAssignmentType *>> matrix_res;
-	for (size_t i = 0; i < lhs.get_num_rows(); ++i)
+	for (int i = 0; i < lhs.get_num_rows(); ++i)
 	{
 		std::vector<BaseAssignmentType *> row;
 
-		for (size_t j = 0; j < lhs.get_num_cols(); j++)
+		for (int j = 0; j < lhs.get_num_cols(); j++)
 		{
 			BaseAssignmentType *res_term = matrix[i][j]->div(&rhs);
 			if (res_term == nullptr) throw Ft_error("Invalid operation");
@@ -83,11 +83,11 @@ Matrix operator / (Matrix lhs, ImaginaryNumber rhs)
 	const std::vector<std::vector<BaseAssignmentType *>> matrix = lhs.get_matrix();
 
 	std::vector<std::vector<BaseAssignmentType *>> matrix_res;
-	for (size_t i = 0; i < lhs.get_num_rows(); ++i)
+	for (int i = 0; i < lhs.get_num_rows(); ++i)
 	{
 		std::vector<BaseAssignmentType *> row;
 
-		for (size_t j = 0; j < lhs.get_num_cols(); j++)
+		for (int j = 0; j < lhs.get_num_cols(); j++)
 		{
 			BaseAssignmentType *res_term = matrix[i][j]->div(&rhs);
 			if (res_term == nullptr) throw Ft_error("Invalid operation");
