+++ srcs/mixed_operators/rational_num_operators.cpp
@@ -80,11 +80,11 @@ Matrix operator * (RationalNumber lhs, Matrix rhs)
 	const std::vector<std::vector<BaseAssignmentType *>> matrix = rhs.get_matrix();
 
 	std::vector<std::vector<BaseAssignmentType *>> matrix_res;
-	for (size_t i = 0; i < rhs.get_num_rows(); ++i)
+	for (int i = 0; i < rhs.get_num_rows(); ++i)
 	{
 		std::vector<BaseAssignmentType *> row;
 
-		for (size_t j = 0; j < rhs.get_num_cols(); j++)
+		for (int j = 0; j < rhs.get_num_cols(); j++)
 		{
 			BaseAssignmentType *res_term = lhs.mult(matrix[i][j]);
 			if (res_term == nullptr) throw Ft_error("Invalid operation");
